+++ tests/blam-tests.lua
@@ -40,12 +40,12 @@ set_callback("command", "OnCommand")
 testFunctions = {}
 
 ------------------------------------------------------------------------------
--- Blam Objects Test Setup
+-- Tag Objects
 ------------------------------------------------------------------------------
 
-testObjects = {}
+testTagObjects = {}
 
-function testObjects:setUp()
+function testTagObjects:setUp()
     self.bipedTagPath = "characters\\cyborg_mp\\cyborg_mp"
     self.uiDefaultProfilesTagPath = "ui\\ui_default_profiles"
     self.assaultRifleWphiTagPath = "weapons\\assault rifle\\assault rifle"
@@ -59,11 +59,7 @@ function testObjects:setUp()
     }
 end
 
-------------------------------------------------------------------------------
--- Tag Objects
-------------------------------------------------------------------------------
-
-function testObjects:testTag()
+function testTagObjects:testTag()
     local cyborgMpTag = blam.getTag(self.bipedTagPath, blam.tagClasses.biped)
     lu.assertNotIsNil(cyborgMpTag, "Cyborg tag must not be nil")
     local bipedTagAddress = get_tag(blam.tagClasses.biped, self.bipedTagPath)
@@ -73,7 +69,7 @@ function testObjects:testTag()
     lu.assertEquals(cyborgMpTag.class, "bipd", "Cyborg tag class type must be bipd")
 end
 
-function testObjects:testTagCollection()
+function testTagObjects:testTagCollection()
     local uiDefaultProfiles = blam.tagCollection(self.uiDefaultProfilesTagPath)
     lu.assertNotIsNil(uiDefaultProfiles, "Tag collection must not be nil")
     lu.assertEquals(uiDefaultProfiles.count, 4, "Tag collection count must be 4")
@@ -81,12 +77,12 @@ function testObjects:testTagCollection()
                     "Tag collection list must match")
 end
 
-function testObjects:testModelAnimations()
+function testTagObjects:testModelAnimations()
     local assaultRifleAnimations = blam.modelAnimations(self.assaultRifleFpAnimsTagPath)
     lu.assertEquals(assaultRifleAnimations.fpAnimationList, self.assaultRifleFpAnimsValues)
 end
 
-function testObjects:testWeaponHudInterface()
+function testTagObjects:testWeaponHudInterface()
     local wphi = blam.weaponHudInterface(self.assaultRifleWphiTagPath)
     lu.assertEquals(wphi.childHud, 3800891673)
     lu.assertEquals(wphi.totalAmmoCutOff, 61)
@@ -112,17 +108,49 @@ end
 ------------------------------------------------------------------------------
 -- Game Objects
 ------------------------------------------------------------------------------
+testObjects = {}
+
+function testObjects:setUp()
+    self.assaultRifleTagPath = "weapons\\assault rifle\\assault rifle"
+end
 
 function testObjects:testBipedObject()
     local bipedObject = blam.biped(get_dynamic_player())
     lu.assertNotIsNil(bipedObject)
     lu.assertNotIsNil(bipedObject.structure)
     lu.assertEquals(bipedObject.invisible, false, "Biped should not have camouflage active")
-    lu.assertEquals(bipedObject.shield, 1)
-    lu.assertEquals(bipedObject.health, 1)
-    lu.assertEquals(bipedObject.isNotDamageable, false)
-    lu.assertEquals(bipedObject.bumpedObjectId, 0)
-    lu.assertIsTrue(blam.isNull(bipedObject.landing))
+    lu.assertEquals(bipedObject.shield, 1, "Biped shield should not depleted")
+    lu.assertEquals(bipedObject.health, 1, "Biped health should not depleted")
+    lu.assertEquals(bipedObject.isNotDamageable, false, "Biped should be damageable")
+    lu.assertEquals(bipedObject.bumpedObjectId, 0, "Biped should not be bumping other objects")
+    if (blam.isNull(bipedObject.vehicleObjectId)) then
+        lu.assertIsTrue(blam.isNull(bipedObject.vehicleObjectId), "Biped should not be inside a vehicle")
+        lu.assertIsTrue(blam.isNull(bipedObject.vehicleSeatIndex), "Biped should not be seated on a vehicle")
+    else
+        lu.assertIsTrue(not blam.isNull(bipedObject.vehicleObjectId), "Biped is inside a vehicle")
+        lu.assertEquals(bipedObject.vehicleSeatIndex, 0, "Biped is seated on seat index 0")
+    end
+    lu.assertIsTrue(blam.isNull(bipedObject.landing), "Biped should not be landing")
+end
+
+------------------------------------------------------------------------------
+-- Data Objects
+------------------------------------------------------------------------------
+
+function testObjects:testPlayerDataObject()
+    local player = blam.player(get_player())
+    lu.assertNotIsNil(player, "Player object must not be nil")
+    lu.assertNotIsNil(player.id, "Player id must not be nil")
+    lu.assertEquals(player.host, 0, "Player should be host")
+    local playerName = blam.readUnicodeString(get_player() + 0x4, true)
+    lu.assertEquals(player.name, playerName, "Player names should be equal")
+    lu.assertEquals(player.team, 0, "Player team should be red")
+    lu.assertNotIsNil(player.objectId, "Player objectId mut not be nil")
+    lu.assertEquals(get_object(player.objectId), get_dynamic_player(), "Player object id address should match address from player object")
+    lu.assertNotIsNil(player.color, 0, "Player color must not be nil")
+    lu.assertEquals(player.index, 0, "Player index should be 0")
+    lu.assertEquals(player.speed, 1, "Player speed should be 1")
+    lu.assertIsTrue((player.ping >= 0), "Player ping should be greater than 0 or equal 0")
 end
 
 -- Mocked arguments and executions for standalone execution and in game execution